fix(web): isolate New Relic timing in Suspense boundary
Cache Components forbids uncached async data outside <Suspense>. The
async root layout awaited the agent connection, blocking prerender of
every route (e.g. /_not-found). Move the browser timing header into a
streamed child component so the layout stays static.
