+++ apps/web/app/layout.tsx
@@ -1,11 +1,10 @@
 import './globals.css';
-import React from 'react';
-import newrelic from 'newrelic';
+import React, { Suspense } from 'react';
 import type { Metadata } from 'next';
 import { Roboto } from 'next/font/google';
-import Script from 'next/script';
 import { Container, Footer, Header } from '~/components/shared';
 import { Toaster } from '~/components/ui/sonner';
+import { NewRelicBrowserTiming } from './newrelic-browser-timing';
 
 const roboto = Roboto({
   display: 'swap',
@@ -17,32 +16,17 @@ export const metadata: Metadata = {
   description: 'A web app that interacts with the Twitch API',
 };
 
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  // @ts-ignore
-  if (newrelic.agent.collector.isConnected() === false) {
-    await new Promise((resolve) => {
-      // @ts-ignore
-      newrelic.agent.on('connected', resolve);
-    });
-  }
-
-  const browserTimingHeader = newrelic.getBrowserTimingHeader({
-    hasToRemoveScriptWrapper: true,
-    allowTransactionlessInjection: true,
-  });
-
   return (
     <html lang='en' className={roboto.className}>
       <head>
-        <Script
-          id='nr-browser-agent'
-          strategy='beforeInteractive'
-          dangerouslySetInnerHTML={{ __html: browserTimingHeader }}
-        />
+        <Suspense fallback={null}>
+          <NewRelicBrowserTiming />
+        </Suspense>
         <link rel='icon' href='/favicon.ico' sizes='any' />
         <link rel='icon' href='/icon.png' type='image/png' sizes='32x32' />
       </head>
